generator_server.py 통합
